pyproject.toml
[project]
name = "frida-tool"
version = "1.0.0"
description = "Toan Nguyen's Hacking Tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.3",
    "mcp>=1.0.0",
    "pyopenssl>=25.1.0",
    "pure-python-adb>=0.3.0.dev0",
]

[project.scripts]
frida-tool = "cli_tool:main"
